Extract product utils into their own file
