+++ app/client/components/cancel/cancellationSummary.tsx
@@ -4,10 +4,8 @@ import { Link } from "@reach/router";
 import React from "react";
 import { cancellationFormatDate } from "../../../shared/dates";
 import { ProductDetail, Subscription } from "../../../shared/productResponse";
-import {
-  hasDeliveryRecordsFlow,
-  ProductType
-} from "../../../shared/productTypes";
+import { ProductType } from "../../../shared/productTypes";
+import { hasDeliveryRecordsFlow } from "../../productUtils";
 import { GenericErrorScreen } from "../genericErrorScreen";
 import { ResubscribeThrasher } from "../resubscribeThrasher";
 import { SupportTheGuardianButton } from "../supportTheGuardianButton";
